Refactor event-card into component, Fixed issue #109
