+++ src/main/webapp/app/entities/event/event.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { connect } from 'react-redux';
 import { Link, RouteComponentProps } from 'react-router-dom';
-import { Container, Row, Col, Card, CardImg, Button } from 'reactstrap';
+import { Container, Row, Col, Card, CardImg, Button, Tooltip } from 'reactstrap';
 // tslint:disable-next-line:no-unused-variable
 import { Translate, TextFormat, getSortState, IPaginationBaseState } from 'react-jhipster';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
@@ -21,6 +21,7 @@ import AuthorizationChecker from 'app/shared/components/authorization-checker/au
 import CCRole from 'app/shared/model/enum/cc-role.enum';
 import EventRole from 'app/shared/model/enum/event-role.enum';
 import { IEvent } from 'app/shared/model/event.model';
+import EventCard from './event-card';
 
 export interface IEventProps extends StateProps, DispatchProps, RouteComponentProps<{ url: string }> {}
 
@@ -30,18 +31,23 @@ export type IEventState = IPaginationBaseState & {
 };
 
 export class Event extends React.Component<IEventProps, IEventState> {
-  state: IEventState = {
-    ...getSortState(this.props.location, ITEMS_PER_PAGE),
-    modalIsOpen: false,
-    eventId: null,
-  };
+  constructor(props: IEventProps) {
+    super(props);
 
-  getTab = () => {
-    const path = this.props.location.search;
-    return path.substring(path.lastIndexOf('?') + 1);
-  };
+    this.state = {
+      ...getSortState(this.props.location, ITEMS_PER_PAGE),
+      modalIsOpen: false,
+      eventId: null,
+    };
 
-  componentDidMount() {
+    this.getTab = this.getTab.bind(this);
+    this.getUpcomingEntities = this.getUpcomingEntities.bind(this);
+    this.getPreviousEntities = this.getPreviousEntities.bind(this);
+    this.openModal = this.openModal.bind(this);
+    this.closeModal = this.closeModal.bind(this);
+  }
+
+  componentDidMount(): void {
     const tab = this.getTab();
     if (tab === 'previous') {
       this.getPreviousEntities();
@@ -50,7 +56,7 @@ export class Event extends React.Component<IEventProps, IEventState> {
     }
   }
 
-  componentDidUpdate(prevProps) {
+  componentDidUpdate(prevProps: IEventProps): void {
     const path = this.props.location.search;
     if (prevProps.location.search !== path) {
       const tab = path.substring(path.lastIndexOf('?') + 1);
@@ -62,26 +68,31 @@ export class Event extends React.Component<IEventProps, IEventState> {
     }
   }
 
-  getUpcomingEntities = () => {
+  getTab(): string {
+    const path = this.props.location.search;
+    return path.substring(path.lastIndexOf('?') + 1);
+  }
+
+  getUpcomingEntities(): void {
     const { activePage, itemsPerPage } = this.state;
     this.props.getUpcomingEntities(activePage - 1, itemsPerPage, `startDate,desc`);
-  };
+  }
 
-  getPreviousEntities = () => {
+  getPreviousEntities(): void {
     const { activePage, itemsPerPage } = this.state;
     this.props.getPreviousEntities(activePage - 1, itemsPerPage, `startDate,desc`);
-  };
+  }
 
-  openModal = eventId => {
+  openModal(eventId: number): void {
     this.setState({ modalIsOpen: true, eventId });
-  };
+  }
 
-  closeModal = () => {
+  closeModal(): void {
     this.setState({ modalIsOpen: false, eventId: null });
-  };
+  }
 
   render() {
-    const { eventList, match, totalItems } = this.props;
+    const { eventList, match } = this.props;
     const { eventId } = this.state;
     const tab = this.getTab();
     return (
@@ -123,53 +134,6 @@ export class Event extends React.Component<IEventProps, IEventState> {
     );
   }
 }
-
-interface IEventCardProps {
-  event: IEvent;
-  toggleModal: (eventId: number) => void;
-}
-
-const EventCard: React.FC<IEventCardProps> = ({ event, toggleModal }) => {
-  const onToggleModal = () => toggleModal(event.id);
-  return (
-    <Card className="p-3 pt-4 event-card">
-      <AuthorizationChecker ccRole={CCRole.ADMIN} eventRole={EventRole.HEAD} eventId={event.id}>
-        <Button color="link" className="option-icon p-0" onClick={onToggleModal}>
-          <FontAwesomeIcon icon={'ellipsis-h'} />
-        </Button>
-      </AuthorizationChecker>
-      <Link to={`/entity/event/${event.id}`}>
-        <Row>
-          <Col xs="4" lg="5" className="pr-0">
-            <CardImg
-              height="100%"
-              width="100%"
-              className="rounded-0"
-              src={event.imageUrl ? event.imageUrl : 'content/images/placeholder.png'}
-              alt={event.fileName}
-            />
-          </Col>
-          <Col xs="8" lg="7">
-            <div className="my-auto">
-              <h4 className="event-title">{event.name}</h4>
-              <p className="mb-0">
-                <Translate contentKey="clubmanagementApp.event.startDate">Start Date</Translate>:{' '}
-                <TextFormat type="date" value={event.startDate} format={APP_DATE_12_ABR_FORMAT} />
-              </p>
-              <p className="mb-0">
-                <Translate contentKey="clubmanagementApp.event.endDate">End Date</Translate>:{' '}
-                <TextFormat type="date" value={event.endDate} format={APP_DATE_12_ABR_FORMAT} />
-              </p>
-              <p className="mb-0">
-                <Translate contentKey="clubmanagementApp.event.venue">Venue</Translate>: {event.venue}
-              </p>
-            </div>
-          </Col>
-        </Row>
-      </Link>
-    </Card>
-  );
-};
 
 const mapStateToProps = ({ event }: IRootState) => ({
   eventList: event.entities,
